# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "BranchCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "branch_id": {
            "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
            "title": "Branch Id",
            "type": "string"
          },
          "content_locale": {
            "default": "zh-CN",
            "maxLength": 16,
            "minLength": 2,
            "pattern": "^[a-z]{2,3}(?:-[A-Z]{2})?$",
            "title": "Content Locale",
            "type": "string"
          },
          "parent_branch_id": {
            "default": "main",
            "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
            "title": "Parent Branch Id",
            "type": "string"
          },
          "source_checkpoint_id": {
            "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
            "title": "Source Checkpoint Id",
            "type": "string"
          }
        },
        "required": [
          "branch_id",
          "source_checkpoint_id"
        ],
        "title": "BranchCreateRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProjectionRequest": {
        "additionalProperties": false,
        "properties": {
          "checkpoint_id": {
            "anyOf": [
              {
                "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkpoint Id"
          }
        },
        "title": "ProjectionRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "SessionToken": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Local story-domain sidecar API",
    "title": "AI Story Evolution Engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/projects/{project_id}/branches": {
      "get": {
        "operationId": "list_branches_projects__project_id__branches_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BranchManifest"
                  },
                  "title": "Response List Branches Projects  Project Id  Branches Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "List Branches",
        "tags": [
          "simulations"
        ]
      },
      "post": {
        "operationId": "create_branch_projects__project_id__branches_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BranchCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BranchManifest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "Create Branch",
        "tags": [
          "simulations"
        ]
      }
    },
    "/projects/{project_id}/branches/compare": {
      "get": {
        "operationId": "compare_branches_projects__project_id__branches_compare_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "left",
            "required": true,
            "schema": {
              "title": "Left",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "right",
            "required": true,
            "schema": {
              "title": "Right",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BranchComparisonResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "Compare Branches",
        "tags": [
          "simulations"
        ]
      }
    },
    "/projects/{project_id}/branches/{branch_id}": {
      "get": {
        "operationId": "get_branch_projects__project_id__branches__branch_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "branch_id",
            "required": true,
            "schema": {
              "title": "Branch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BranchManifest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "Get Branch",
        "tags": [
          "simulations"
        ]
      }
    },
...
    "/projects/{project_id}/branches/{branch_id}/projection": {
      "post": {
        "operationId": "rebuild_projection_projects__project_id__branches__branch_id__projection_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "branch_id",
            "required": true,
            "schema": {
              "title": "Branch Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "Rebuild Projection",
        "tags": [
          "simulations"
        ]
      }
    },
    "/projects/{project_id}/branches/{branch_id}/rollback": {
      "post": {
        "operationId": "rollback_branch_projects__project_id__branches__branch_id__rollback_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "branch_id",
            "required": true,
            "schema": {
              "title": "Branch Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BranchRollbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BranchManifest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "Rollback Branch",
        "tags": [
          "simulations"
        ]
      }
    },
    "/projects/{project_id}/branches/{branch_id}/simulation-trace": {
      "get": {
        "operationId": "list_simulation_trace_projects__project_id__branches__branch_id__simulation_trace_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "branch_id",
            "required": true,
            "schema": {
              "title": "Branch Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after_step",
            "required": false,
            "schema": {
              "default": -1,
              "title": "After Step",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SimulationLogRecord"
                  },
                  "title": "Response List Simulation Trace Projects  Project Id  Branches  Branch Id  Simulation Trace Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "List Simulation Trace",
        "tags": [
          "simulations"
        ]
      }
    },
...
    "/projects/{project_id}/simulation-events": {
      "get": {
        "operationId": "list_simulation_events_projects__project_id__simulation_events_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after_sequence",
            "required": false,
            "schema": {
              "default": 0,
              "title": "After Sequence",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EngineEvent"
                  },
                  "title": "Response List Simulation Events Projects  Project Id  Simulation Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "List Simulation Events",
        "tags": [
          "simulations"
        ]
      }
    },
...
  }
}